+++ src/application/UserApplicationService.ts
@@ -1,29 +1,18 @@
 import UserId from '../domain/models/User/UserId/UserId';
-import { TODO } from '../util/Util';
-import InMemoryUserRepository from '../inMemory/InMemoryUserRepository';
+import UserRepository from '../infrastructure/UserRepository';
 import { IUserRepository } from '../domain/models/User/IUserRepository';
-import { BirthdayProps } from '../domain/models/User/Profile/Birthday';
 import { UserDataModel } from '../infrastructure/UserDataModel';
 import { UserFactory } from '../domain/models/User/UserFactory';
 
 // note ここにロジックは書かない。追加のロジックが必要になったらdomain model, domain service に書こう。
 // 引数を受け取って new Hoge() するとかならOK
 export default class UserApplicationService {
-  static readonly userRepository: IUserRepository = new InMemoryUserRepository();
+  static readonly userRepository: IUserRepository = new UserRepository();
 
   static readonly userFactory = new UserFactory();
 
   static getUserIdFromLocalStorage(): UserId | null {
-    return this.userRepository.getUserIdFromLocalStorage();
-  }
-
-  static toInstanceUserId(userIdString: string): UserId {
-    return this.userRepository.toInstanceUserId(userIdString);
-  }
-
-  // todo login 中のユーザーの情報しかもってこれないようにする
-  static findUserByUserId(userId: UserId): TODO<'User'> {
-    return this.userRepository.getUserByUserId(userId);
+    return UserApplicationService.userRepository.getUserIdFromLocalStorage();
   }
 
   static async getCurrentUser(userId: UserId): Promise<UserDataModel> {
@@ -32,41 +21,34 @@ export default class UserApplicationService {
       .catch((e) => e);
     const userJson = await userData.json();
     const user = this.userRepository.toInstance(userJson);
-    const userDataModel = this.userFactory.createUserDataModel(user);
 
-    return userDataModel;
+    return this.userFactory.createUserDataModel(user);
   }
 
-  // todo パラメータごとにsaveメソッドを分けるかどうかはバックエンドと相談(とりあえずわけずにいく)
-  static updateUserName(userId: UserId, userNameString: string): void {
-    const user = UserApplicationService.findUserByUserId(userId);
-    const updatedUser = user.updateUserName(userNameString);
-    UserApplicationService.userRepository.save(updatedUser);
-  }
-
-  static updateBio(userId: UserId, bioString: string): void {
-    const user = UserApplicationService.findUserByUserId(userId);
-    const updatedUser = user.updateBio(bioString);
-    UserApplicationService.userRepository.save(updatedUser);
-  }
-
-  static updateWebsite(userId: UserId, websiteString: string): void {
-    const user = UserApplicationService.findUserByUserId(userId);
-    const updatedUser = user.updateWebsite(websiteString);
-    UserApplicationService.userRepository.save(updatedUser);
-  }
-
-  static updateUserLocation(userId: UserId, userLocationString: string): void {
-    const user = UserApplicationService.findUserByUserId(userId);
-    const updatedUser = user.updateUserLocation(userLocationString);
-    UserApplicationService.userRepository.save(updatedUser);
-  }
-
-  static updateBirthday(userId: UserId, birthdayProps: BirthdayProps): void {
-    const user = UserApplicationService.findUserByUserId(userId);
-    const updatedUser = user.updateBirthday(birthdayProps);
-    UserApplicationService.userRepository.save(updatedUser);
-  }
+  // static updateUserName(user: IUser, userNameString: string): void {
+  //   const updatedUser = user.updateUserName(userNameString);
+  //   UserApplicationService.userRepository.save(updatedUser);
+  // }
+  //
+  // static updateBio(user: IUser, bioString: string): void {
+  //   const updatedUser = user.updateBio(bioString);
+  //   UserApplicationService.userRepository.save(updatedUser);
+  // }
+  //
+  // static updateWebsite(user: IUser, websiteString: string): void {
+  //   const updatedUser = user.updateWebsite(websiteString);
+  //   UserApplicationService.userRepository.save(updatedUser);
+  // }
+  //
+  // static updateUserLocation(user: IUser, userLocationString: string): void {
+  //   const updatedUser = user.updateUserLocation(userLocationString);
+  //   UserApplicationService.userRepository.save(updatedUser);
+  // }
+  //
+  // static updateBirthday(user: IUser, birthdayProps: BirthdayProps): void {
+  //   const updatedUser = user.updateBirthday(birthdayProps);
+  //   UserApplicationService.userRepository.save(updatedUser);
+  // }
 
   static isAuthorized(screenName: string, password: string): boolean {
     return this.userRepository.isAuthorized(screenName, password);
