+++ src/domain/models/User/IUserRepository.ts
@@ -1,12 +1,11 @@
 import UserId from './UserId/UserId';
-import { IUser } from './IUser';
+// import { IUser } from './IUser';
 import { User } from './User';
 // eslint-disable-next-line import/no-cycle
-import { userFull } from '../../../inMemory/InMemoryUserRepository';
+import { userFull } from '../../../infrastructure/UserRepository';
 
 export interface IUserRepository {
-  getUserByUserId(userId: UserId): IUser;
-  save(user: IUser): void;
+  // save(user: IUser): void;
   isAuthorized(screenName: string, password: string): boolean;
   setUserIdToLocalStorage(userId: UserId): void;
   getUserIdFromLocalStorage(): UserId | null;
